feat(docker-compose.yaml): add Redis and PostgreSQL services to the docker-compose file
The Redis service is added to provide a Redis server with version 7.2.4-alpine. It is exposed on port 6379 and set to restart always. The Redis server is configured with a required password "autogen_api".

The PostgreSQL service is added to provide a PostgreSQL database server with version 16.2-alpine3.19. It is exposed on port 5431 and set to restart always. The database is configured with the default database name "postgres", username "postgres", and password "postgres". The data directory is set to "/var/lib/postgresql/data".