chore: Update Docker image tags and add Node.js installation
This commit updates the Docker image tags and adds the installation of Node.js to the Dockerfile.base. The Docker image tags are updated to ensure the latest versions are used, and Node.js is installed using the official NodeSource setup script. This change is made to keep the dependencies up to date and enable the use of Node.js in the project.