Chore/update release drafter (#92)
* chore: Update permissions in GitHub workflows

This commit updates the permissions in the GitHub workflows to use the "write-all" permission instead of specifying individual permissions for contents, pull-requests, and issues. This change simplifies the workflow configuration and ensures that all necessary permissions are granted for the workflows to function properly.

* chore: Update GitHub workflows permissions and remove unused code

This commit updates the GitHub workflows permissions to use the "write-all" permission instead of specifying individual permissions for contents, pull-requests, and issues. This change simplifies the workflow configuration and ensures that all necessary permissions are granted for the workflows to function properly.

Additionally, unused code and configuration files are removed to improve code cleanliness and maintainability.

* chore: Update GitHub workflows permissions and remove unused code

This commit updates the GitHub workflows permissions to use the "write-all" permission instead of specifying individual permissions for contents, pull-requests, and issues. Additionally, unused code and configuration files are removed to improve code cleanliness and maintainability.